Modificaciones por problemas del merge
