+++ API/Infrastructure/Repositorios/PrestamoRepositorio.cs
@@ -13,10 +13,11 @@ namespace Infrastructure.Repositorios
     {
         public PrestamoRepositorio(AppDbContext context) : base(context)
         {
+            
         }
 		public override async Task<IEnumerable<Prestamo>> ObtenerTodosAsincrono()
 		{
 			return await base.dbSet.Include(x => x.Estado).Include(x => x.cliente).Include(x => x.plazo).ToListAsync();
-		
+        }
 	}
 }
