+++ API/Services/Servicios/PrestamoServicio.cs
@@ -2,6 +2,9 @@ using Core.Entidades;
 using Core.Interfaces;
 using Core.Interfaces.Servicios;
 using Core.Respuestas;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+using Services.Validadores;
+using static System.Runtime.InteropServices.JavaScript.JSType;
 
 namespace Services.Servicios
 {
@@ -17,10 +20,22 @@ namespace Services.Servicios
             throw new NotImplementedException();
         }
 
-        public Task<Respuesta<Prestamo>> Agregar(Prestamo nuevaEntitidad)
-        {
-            throw new NotImplementedException();
-        }
+        public async Task<Respuesta<Prestamo>> Agregar(Prestamo nuevaEntitidad)
+		{
+			PrestamoValidador validador = new();
+
+			var resultadoValidacion = await validador.ValidateAsync(nuevaEntitidad);
+
+			if (!resultadoValidacion.IsValid)
+			{
+				throw new ArgumentException(resultadoValidacion.Errors[0].ErrorMessage.ToString());
+			}
+
+			var entidadagregada = await _unidadDeTrabajo.PrestamoRepostorio.AgregarAsincrono(nuevaEntitidad);
+			await _unidadDeTrabajo.CommitAsync();
+
+			return new Respuesta<Prestamo> { Ok = true, Mensaje = "Prestamo creado con éxito", Datos = entidadagregada };
+		}
 
         public Task<Respuesta<Prestamo>> ObternerPorIdAsincrono(int id)
         {
@@ -103,5 +118,87 @@ namespace Services.Servicios
                 throw e;
             }
         }
+
+        public async Task<Respuesta<Prestamo>> ConsultarPrestamoValidado(int idUsuarioSesion, int idPrestamo)
+		{
+			if (idUsuarioSesion == null || idUsuarioSesion == 0)
+			{
+				throw new ArgumentException("No se ha insertado el id del usuario de la sesión activa.");
+			}
+
+			Prestamo prestamo = await _unidadDeTrabajo.PrestamoRepostorio.ObtenerPorIdAsincrono(idPrestamo);
+
+			Usuario usuario = await _unidadDeTrabajo.UsuarioRepositorio.ObtenerPorIdAsincrono(idUsuarioSesion);
+
+			if(prestamo.IdCliente != usuario.ClienteId)
+			{
+				return new Respuesta<Prestamo> { Ok = false, Mensaje = "Consulta inválida. No se puede consultar un préstamo que no pertenezca al usuario actual", Datos = null };
+			}
+			else
+			{
+				return new Respuesta<Prestamo> { Ok = true, Mensaje = "Préstamo obtenido con éxito", Datos = prestamo };
+
+			}
+		}
+
+		public async Task<Respuesta<Prestamo>> SolicitarPrestamo(int idUsuarioSesion, ModeloSolicitudPrestamo modeloSolicitudPrestamo)
+		{
+			if (idUsuarioSesion == null || idUsuarioSesion == 0)
+			{
+				throw new ArgumentException("No se ha insertado el id del usuario de la sesión activa.");
+			}
+
+			Usuario usuario = await _unidadDeTrabajo.UsuarioRepositorio.ObtenerPorIdAsincrono(idUsuarioSesion);
+
+			//Aquí validación de que modeloSolicitudPrestamo.DocumentoIdentificacionPersonal y modeloSolicitudPrestamo.DocumentoComprobanteIngresos
+			//sean documentos válidos.
+
+			if (modeloSolicitudPrestamo.MontoTotalDeseado > (modeloSolicitudPrestamo.SueldoBasicoDelSolicitante * 3))
+			{
+				return new Respuesta<Prestamo> { Ok = false, Mensaje = "El monto solicitado supera el equivalente a tres sueldos básicos del solicitante.", Datos = null };
+			}
+
+			var PlazoIdeal = await _unidadDeTrabajo.PlazoRepositorio.ConsultarPlazoIdeal(modeloSolicitudPrestamo.NumeroCuotasDeseadas);
+
+			//hacer metodo repositorio para buscar minimo cuotas y maximo cuotas para indicar el error abajo.
+
+			if (PlazoIdeal == null)
+			{
+				return new Respuesta<Prestamo> { Ok = false, Mensaje = "El número de cuotas solicitadas es inválido, probar con otro número", Datos = null };
+			}
+
+			double cuotaMensual = (modeloSolicitudPrestamo.MontoTotalDeseado * PlazoIdeal.Porcentaje) / Math.Pow((1 - (1 + PlazoIdeal.Porcentaje)), (-1*modeloSolicitudPrestamo.NumeroCuotasDeseadas));
+
+
+			var prestamoAgregado = await _unidadDeTrabajo.PrestamoRepostorio.AgregarAsincrono(new Prestamo
+			{
+				Id = 0,
+				NumeroCuotas = modeloSolicitudPrestamo.NumeroCuotasDeseadas,
+				MontoTotal = modeloSolicitudPrestamo.MontoTotalDeseado,
+				CuotaMensual = cuotaMensual,
+				Fecha = DateTime.Now,
+				IdEstado = 1,
+				IdCliente = usuario.ClienteId,
+				IdPlazo = PlazoIdeal.Id
+			});
+			//suponiendo que idestado 1 sea un estado inicial como, en proceso de pago
+
+
+			//crear cuotas
+			//for modeloSolicitudPrestamo.NumeroCuotasDeseadas
+			//await unidad de trabajo cuotarepositori.agregar(id = 0, prestamoid = pretamoagregado.id, fecha = DateTime.Now.AddMonths(i+1);)
+
+
+
+
+
+
+			await _unidadDeTrabajo.CommitAsync();
+
+			//aqui va agregar documentos relacionados al id de lo de arriba
+
+			return new Respuesta<Prestamo> { Ok = true, Mensaje = "Prestamo creado con éxito", Datos = prestamoAgregado };
+
+		}
     }
 }
